+++ Changelog.md
@@ -7,6 +7,7 @@
 * Use more of gobject-introspection's test libraries for testing
 * Simplify constructor overrides: Custom initializers will no longer be
   overwritten when setting up the corresponding constructor.
+* Override GObject::Object.new to not require a GType argument
 
 ## 0.8.3 / 2015-11-13